# perf: skip visiting if it's out of range