--- a/crates/turbopack-ecmascript/src/path_visitor.rs
+++ b/crates/turbopack-ecmascript/src/path_visitor.rs
@@ -29,13 +29,19 @@ fn find_range<'a, 'b>(
     index: usize,
 ) -> Option<&'b [(&'a AstPath, &'a dyn VisitorFactory)]> {
     // Precondition: visitors is never empty
-    if visitors.first().unwrap().0[index] > *kind {
+    if visitors.first().unwrap().0[index] > *kind || visitors.last().unwrap().0[index] < *kind {
         // Fast path: If ast path of the first visitor is already out of range, then we
         // can skip the whole visit.
         return None;
     }
 
-    let start = visitors.partition_point(|(path, _)| path[index] < *kind);
+    let start = if visitors.first().unwrap.0[index] == *kind {
+        // Fast path: It looks like the whole range is selected
+        0
+    } else {
+        visitors.partition_point(|(path, _)| path[index] < *kind)
+    };
+
     if start >= visitors.len() {
         return None;
     }