@sokra 2023-02-13
## Benchmark for 709736a

| Test | Base         | PR               | % | Significant % |
|------|--------------|------------------|---|--------------|
| bench_startup/Turbopack RCC/1000 modules | 2636.08ms ± 4.17ms | **2670.38ms ± 7.63ms** | +1.30% | +0.41% |
  
<details>
  <summary>Click to view full benchmark</summary>
  
| Test | Base         | PR               | % | Significant % |
|------|--------------|------------------|---|--------------|
| bench_hmr_to_commit/Turbopack CSR/1000 modules | 9914.92µs ± 89.24µs | 10.16ms ± 0.10ms | +2.49% |  |
| bench_hmr_to_commit/Turbopack RCC/1000 modules | 12.42ms ± 0.25ms | 12.58ms ± 0.35ms | +1.34% |  |
| bench_hmr_to_commit/Turbopack RSC/1000 modules | 540.67ms ± 2.64ms | 539.90ms ± 3.47ms | -0.14% |  |
| bench_hmr_to_commit/Turbopack SSR/1000 modules | 10.17ms ± 0.10ms | 10.25ms ± 0.07ms | +0.79% |  |
| bench_hmr_to_eval/Turbopack CSR/1000 modules | 8953.30µs ± 66.66µs | 9060.92µs ± 94.88µs | +1.20% |  |
| bench_hmr_to_eval/Turbopack RCC/1000 modules | 12.29ms ± 0.34ms | 11.90ms ± 0.38ms | -3.16% |  |
| bench_hmr_to_eval/Turbopack SSR/1000 modules | 9059.42µs ± 91.76µs | 9106.22µs ± 43.67µs | +0.52% |  |
| bench_hydration/Turbopack RCC/1000 modules | 4410.91ms ± 14.30ms | 4445.47ms ± 14.91ms | +0.78% |  |
| bench_hydration/Turbopack RSC/1000 modules | 3988.58ms ± 21.76ms | 3999.80ms ± 14.67ms | +0.28% |  |
| bench_hydration/Turbopack SSR/1000 modules | 3835.85ms ± 14.59ms | 3877.63ms ± 17.86ms | +1.09% |  |
| bench_startup/Turbopack CSR/1000 modules | 2908.43ms ± 10.43ms | 2875.43ms ± 10.79ms | -1.13% |  |
| bench_startup/Turbopack RCC/1000 modules | 2636.08ms ± 4.17ms | **2670.38ms ± 7.63ms** | +1.30% | +0.41% |
| bench_startup/Turbopack RSC/1000 modules | 2578.77ms ± 8.70ms | 2553.38ms ± 6.72ms | -0.98% |  |
| bench_startup/Turbopack SSR/1000 modules | 2188.13ms ± 6.12ms | 2194.88ms ± 3.13ms | +0.31% |  |
  
</details>